+++ packages/ui/src/form/primitives/searchable-chip-select/searchable-chip-select.tsx
@@ -6,15 +6,9 @@ import defenseStyles from '../../../utils/css/global-css-defense.module.css';
 import focusStyles from '../../../utils/css/focus.module.scss';
 import * as Combobox from '../combobox';
 import { InputLayout } from '../input-layout';
+import { SearchableCollection } from '../searchable-collection';
 import styles from './style.module.css';
-import { warnSearchableChipSelectProps } from './dev-warnings';
-import type { Item, ItemGroup, SearchableChipSelectProps } from './types';
-import {
-	findCreatableItem,
-	isItem,
-	normalizeRootItems,
-	shouldSkipCollectionEntry,
-} from './types';
+import type { Item, SearchableChipSelectProps } from './types';
 
 /**
  * A low-level primitive for a searchable multi-selection field with chips, with
@@ -43,14 +37,11 @@ export const SearchableChipSelect = forwardRef<
 	},
 	ref
 ) {
-	warnSearchableChipSelectProps( items, children );
-
-	const creatableItem = findCreatableItem( items );
-	const comboboxItems = normalizeRootItems( items );
-
 	return (
-		<Combobox.Root
-			items={ comboboxItems }
+		<SearchableCollection.Root
+			name="SearchableChipSelect"
+			items={ items }
+			collection={ children }
 			multiple
 			disabled={ disabled }
 			{ ...restProps }
@@ -123,53 +114,8 @@ export const SearchableChipSelect = forwardRef<
 			</Combobox.InputGroup>
 
 			<Combobox.Popup width={ popupWidth }>
-				<Combobox.Empty>{ emptyContent }</Combobox.Empty>
-				<Combobox.List>
-					<Combobox.ListBody>
-						<Combobox.Collection>
-							{ ( entry: Item | ItemGroup, ...args ) => {
-								if (
-									shouldSkipCollectionEntry(
-										entry,
-										creatableItem
-									)
-								) {
-									return null;
-								}
-
-								if ( children ) {
-									return children( entry, ...args );
-								}
-
-								if ( ! isItem( entry ) ) {
-									return null;
-								}
-
-								return (
-									<Combobox.Item
-										key={ entry.value }
-										value={ entry }
-										disabled={ entry.disabled }
-									>
-										{ entry.label }
-									</Combobox.Item>
-								);
-							} }
-						</Combobox.Collection>
-					</Combobox.ListBody>
-					{ creatableItem && (
-						<Combobox.ListFooter>
-							<Combobox.Item
-								variant="creatable"
-								value={ creatableItem }
-								disabled={ creatableItem.disabled }
-							>
-								{ creatableItem.label }
-							</Combobox.Item>
-						</Combobox.ListFooter>
-					) }
-				</Combobox.List>
+				<SearchableCollection.List emptyContent={ emptyContent } />
 			</Combobox.Popup>
-		</Combobox.Root>
+		</SearchableCollection.Root>
 	);
 } );